moving the try except block from hiboutik_router to hiboutik_api_class + start creating the route of the user

+++ project/back/src/hiboutik/hitboutik_api_class.py
@@ -1,6 +1,7 @@
 import requests
 import base64
 import os
+from fastapi.exceptions import HTTPException
 
 class HiboutikApi:
     _username = os.environ.get('HIBOUTIK_API_USERNAME')
@@ -10,29 +11,33 @@ class HiboutikApi:
     @classmethod
     def __get_auth_header(cls):
         if cls._username is None or cls._token is None or cls._url is None:
-            raise Exception('HiboutikApi: missing environment variables')
+            raise HTTPException(status_code=500, detail="Internal Server Error")
         base64string = base64.b64encode(bytes('%s:%s' % (cls._username, cls._token), 'ascii'))
         header = "Basic %s" % base64string.decode('utf-8')
         return {'Authorization': header}
 
     @classmethod
     def get(cls, path, params=None):
-        r = requests.get(cls._url + path, params=params, headers=cls.__get_auth_header())
+        if (header := cls.__get_auth_header()) is not None:
+            r = requests.get(cls._url + path, params=params, headers=cls.__get_auth_header())
         return r
 
     @classmethod
     def post(cls, path, data):
-        r = requests.post(cls._url + path, data=data, headers=cls.__get_auth_header())
+        if (header := cls.__get_auth_header()) is not None:
+            r = requests.post(cls._url + path, data=data, headers=cls.__get_auth_header())
         return r
     
     @classmethod
     def put(cls, path, data):
-        r = requests.put(cls._url + path, data=data, headers=cls.__get_auth_header())
+        if (header := cls.__get_auth_header()) is not None:
+            r = requests.put(cls._url + path, data=data, headers=cls.__get_auth_header())
         return r
 
     @classmethod
     def delete(cls, path):
-        r = requests.delete(cls._url + path, headers=cls.__get_auth_header())
+        if (header := cls.__get_auth_header()) is not None:
+            r = requests.delete(cls._url + path, headers=cls.__get_auth_header())
         return r
     
     # Methods not implemented :
